End-to-end desktop wallet testing (#9)
Added testing framework (shamelessly cribbed from the Unit-e Python
functional testing framework), which starts the required Unit-e
daemons in the background. Also, implemented a new environment ('e2e'),
which sets the base RPC port for communicated with those daemons. The
end-to-end tests should now be run as `ng e2e --env=e2e` (a bit redundant,
but whatever.)

Added new yarn command, `travis:e2e`, which starts the end-to-end
tests. Configured Travis to download the Unit-e release and unpack it
for the testing framework to find. The path to the binary is governed
by the `UNITED_BINARY` environment variable.

Signed-off-by: Mihai Ciumeica <mihai@thirdhash.com>