Add replace-by-fee support
[Replace-by-fee](https://github.com/bitcoin/bips/blob/master/bip-0125.mediawiki)
allows nodes to increase the fee on a transaction, if it was initially too small
to commit it to the blockchain, and the transaction was stuck in the mempool.

This commit allows the user to create replaceable transactions, as well as "bump"
the fee for any of his replaceable transactions which are not confirmed yet.

Signed-off-by: Mihai Ciumeica <mihai@thirdhash.com>