Allow sending to multiple recipients
This commit allows configuring multiple outputs for one transaction.
Each output specifies the address, amount, and whether the transaction
fee should be subtracted from it. See the `TransactionBuilder` and
`TransactionOutput` model classes for the implementation.

On the UI side, each output is an instance of the `SendOutputComponent` class,
which contains controls that were previously in `SendComponent`.

Signed-off-by: Mihai Ciumeica <mihai@thirdhash.com>