Export transaction history
Added a new button on the History page, which triggers a prompt for the
user to choose an export location, retrieves the list of transactions from
the daemon, and generates a CSV file with those transactions.

The CSV follows the format given by the Qt wallet and includes the following
columns: Confirmed, Date, Type, Label, Address, Amount, ID. The only difference
is that the Qt wallet decomposes wallet transactions into several rows, for
display and export purposes, with one row corresponding to a Tx output. Since
the Electron wallet doesn't do it on the History page, I chose not to decompose
transactions in the CSV export either, for consistency.

Signed-off-by: Mihai Ciumeica <mihai@thirdhash.com>