Allow user to ignore remotely staked coins when sending funds
This commit adds a checkbox to the coin selection component which allows
the user to ignore any coins staked on other nodes, when creating a transaction.

Changing the checkbox will update the available balance display accordingly.
If the user chooses to select individual coins manually, the checkbox is disabled
and has no effect.

Signed-off-by: Mihai Ciumeica <mihai@thirdhash.com>