Allow user to set custom path to Unit-e daemon
In production, the daemon will be downloaded and unpacked automatically,
however, during development, the user might want to launch a version of
the daemon which they've built themselves.

This commit enables a new command-line argument, `--customdaemon=xxx`,
which lets the developer set a custom path to the Unit-e binary. This
binary will be launched automatically by the wallet on startup.

Signed-off-by: Mihai Ciumeica <mihai@thirdhash.com>