Refactor and add unit tests for the update downloads
On startup, the desktop wallet tries to download a JSON file which
contains information on the latest released version of Unit-e, then
compares the information in the file to the binaries already existing on
disk, and downloads the updated files if needed.

For the structure of the JSON file, see `clientBinaries.json` in this
repository.

This commit refactors the update code from a promise hell into a more
testable async/await style and adds unit tests for it.

Signed-off-by: Mihai Ciumeica <mihai@thirdhash.com>