Allow running the wallet without the Angular server (#68)
This commit introduces a new command-line option for the wallet,
`--dev`, which both opens the Chrome development tools on startup,
and instructs Electron to loads the assets from the Angular development
server, instead of the filesystem. This has the advantage of allowing
live reload when the HTML or script code is changed on disk.

If the `--dev` option is omitted (e.g. when the wallet is started using
`yarn start`), starting the server is no longer necessary.

Signed-off-by: Mihai Ciumeica <mihai@thirdhash.com>